--HG-- branch : xerial-core extra : convert_revision : svn%3Aae02f08e-27ec-0310-ae8c-8ba02fe2eafd/XerialJ/trunk/xerial-core%403338

+++ xerial-core/src/main/java/org/xerial/silk/SilkPullParser.java
@@ -180,7 +180,7 @@ public class SilkPullParser
         return;
     }
 
-    public int getNumReadLine()
+    public long getNumReadLine()
     {
         return parser.getNumReadLine();
     }
